Refactoring enums next() and prev() methods

+++ src/enums.rs
@@ -1,8 +1,8 @@
 use std::convert::TryFrom;
 
-pub fn next<S: Into<u8> + TryFrom<u8>>(val: S, count: u8) -> S {
+pub fn next<S: Into<u8> + TryFrom<u8>>(val: S, count: usize) -> S {
     let mut i: u8 = val.into();
-    if i < count - 1 {
+    if (i as usize) < count - 1 {
         i += 1;
     } else {
         i = 0;
@@ -10,12 +10,12 @@ pub fn next<S: Into<u8> + TryFrom<u8>>(val: S, count: u8) -> S {
     S::try_from(i).ok().unwrap()
 }
 
-pub fn prev<S: Into<u8> + TryFrom<u8>>(val: S, count: u8) -> S {
+pub fn prev<S: Into<u8> + TryFrom<u8>>(val: S, count: usize) -> S {
     let mut i: u8 = val.into();
     if i > 0 {
         i -= 1;
     } else {
-        i = count - 1;
+        i = count as u8 - 1;
     }
     S::try_from(i).ok().unwrap()
 }
